// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMINISTRADOR: 'ADMINISTRADOR',
  RECEPCIONISTA: 'RECEPCIONISTA',
  ESTUDIANTE: 'ESTUDIANTE',
  PROFESOR: 'PROFESOR'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const EstadoUsuario = {
  PENDIENTE_APROBACION: 'PENDIENTE_APROBACION',
  ACTIVO: 'ACTIVO',
  SUSPENDIDO_MORA: 'SUSPENDIDO_MORA',
  INACTIVO: 'INACTIVO'
} as const

export type EstadoUsuario = (typeof EstadoUsuario)[keyof typeof EstadoUsuario]


export const Dias = {
  LUNES: 'LUNES',
  MARTES: 'MARTES',
  MIERCOLES: 'MIERCOLES',
  JUEVES: 'JUEVES',
  VIERNES: 'VIERNES'
} as const

export type Dias = (typeof Dias)[keyof typeof Dias]


export const EstadoGrupo = {
  ABIERTO: 'ABIERTO',
  CANCELADO: 'CANCELADO',
  CERRADO: 'CERRADO'
} as const

export type EstadoGrupo = (typeof EstadoGrupo)[keyof typeof EstadoGrupo]


export const RazonPago = {
  MATRICULA: 'MATRICULA',
  MENSUALIDAD: 'MENSUALIDAD',
  OTRO: 'OTRO'
} as const

export type RazonPago = (typeof RazonPago)[keyof typeof RazonPago]


export const EstadoDeuda = {
  PENDIENTE: 'PENDIENTE',
  PAGADO: 'PAGADO',
  CANCELADA: 'CANCELADA',
  VENCIDO: 'VENCIDO'
} as const

export type EstadoDeuda = (typeof EstadoDeuda)[keyof typeof EstadoDeuda]


export const MetodoPago = {
  TRANSFERENCIA: 'TRANSFERENCIA',
  CAJA: 'CAJA',
  PASARELA_EN_LINEA: 'PASARELA_EN_LINEA'
} as const

export type MetodoPago = (typeof MetodoPago)[keyof typeof MetodoPago]


export const EstadoInscripcion = {
  INSCRITO: 'INSCRITO',
  RETIRADO: 'RETIRADO'
} as const

export type EstadoInscripcion = (typeof EstadoInscripcion)[keyof typeof EstadoInscripcion]


export const EstadoPeriodo = {
  PREPARADO: 'PREPARADO',
  ACTIVO: 'ACTIVO',
  CERRADO: 'CERRADO'
} as const

export type EstadoPeriodo = (typeof EstadoPeriodo)[keyof typeof EstadoPeriodo]


export const EstadoPago = {
  ACEPTADO: 'ACEPTADO',
  RECHAZADO: 'RECHAZADO',
  PENDIENTE: 'PENDIENTE'
} as const

export type EstadoPago = (typeof EstadoPago)[keyof typeof EstadoPago]